+++ src/TestSuitePortal/AbstractTestCase.php
@@ -9,9 +9,13 @@ use Heptacom\HeptaConnect\Core\Component\Composer\PackageConfigurationLoader;
 use Heptacom\HeptaConnect\Core\Configuration\ConfigurationService;
 use Heptacom\HeptaConnect\Core\Configuration\Contract\ConfigurationServiceInterface;
 use Heptacom\HeptaConnect\Core\Configuration\Contract\PortalNodeConfigurationProcessorInterface;
+use Heptacom\HeptaConnect\Core\File\Filesystem\RewritePathStreamWrapper;
+use Heptacom\HeptaConnect\Core\File\Filesystem\StreamUriSchemePathConverter;
 use Heptacom\HeptaConnect\Core\Portal\ComposerPortalLoader;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PortalRegistryInterface;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PortalStackServiceContainerBuilderInterface;
+use Heptacom\HeptaConnect\Core\Portal\File\Filesystem\Contract\FilesystemFactoryInterface;
+use Heptacom\HeptaConnect\Core\Portal\File\Filesystem\Filesystem;
 use Heptacom\HeptaConnect\Core\Portal\PortalFactory;
 use Heptacom\HeptaConnect\Core\Portal\PortalRegistry;
 use Heptacom\HeptaConnect\Core\Portal\PortalStackServiceContainerBuilder;
@@ -35,6 +39,7 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeConfiguration\P
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeConfiguration\PortalNodeConfigurationSetActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
 use Heptacom\HeptaConnect\Storage\Base\PreviewPortalNodeKey;
+use Http\Discovery\Psr17FactoryDiscovery;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
 use Symfony\Component\Cache\Adapter\NullAdapter;
@@ -74,7 +79,7 @@ abstract class AbstractTestCase extends TestCase
         $composerPortalLoader = $this->getPortalLoader();
         $portal = $composerPortalLoader->getPortals()->first();
 
-        return new PreviewPortalNodeKey($portal::class);
+        return new PreviewPortalNodeKey($portal::class());
     }
 
     protected function getPortalLoader(): ComposerPortalLoader
@@ -121,6 +126,7 @@ abstract class AbstractTestCase extends TestCase
             $services[PublisherInterface::class] ?? $this->createMock(PublisherInterface::class),
             $services[HttpHandlerUrlProviderFactoryInterface::class] ?? $httpHandlerUrlProviderFactory,
             $services[RequestStorageContract::class] ?? $this->createMock(RequestStorageContract::class),
+            $services[FilesystemFactoryInterface::class] ?? $this->createFilesystemFactoryForDirectory($this->createRandomTemporaryDirectory()),
         );
 
         $builder->setDirectEmissionFlow(
@@ -155,5 +161,44 @@ abstract class AbstractTestCase extends TestCase
     protected function getProfilerFactory(): ProfilerFactoryContract
     {
         return $this->createMock(ProfilerFactoryContract::class);
+    }
+
+    protected function createFilesystemFactoryForDirectory(string $rootDirectory): FilesystemFactoryInterface
+    {
+        $result = $this->createMock(FilesystemFactoryInterface::class);
+        $result->method('create')->willReturnCallback(static function () use ($rootDirectory) {
+            if (\in_array('portal-node', \stream_get_wrappers(), true)) {
+                \stream_wrapper_unregister('portal-node');
+            }
+
+            if (\in_array('portal-test-local', \stream_get_wrappers(), true)) {
+                \stream_wrapper_unregister('portal-test-local');
+            }
+
+            \FilesystemStreamWrapper::register('portal-test-local', $rootDirectory);
+
+            \stream_wrapper_register('portal-node', RewritePathStreamWrapper::class);
+            \stream_context_set_default([
+                'portal-node' => [
+                    'protocol' => [
+                        'set' => 'portal-test-local',
+                    ],
+                ],
+            ]);
+
+            return new Filesystem(new StreamUriSchemePathConverter(Psr17FactoryDiscovery::findUriFactory(), 'portal-node'));
+        });
+
+        return $result;
+    }
+
+    protected function createRandomTemporaryDirectory(): string
+    {
+        $tempDir = \rtrim(\sys_get_temp_dir(), '/\\');
+        $result = $tempDir . \DIRECTORY_SEPARATOR . 'heptaconnect-test-suite-portal' . \DIRECTORY_SEPARATOR . \bin2hex(\random_bytes(32));
+
+        \mkdir($result, 0777, true);
+
+        return $result;
     }
 }